# frontend/openapi.json
{
  "components": {
...
    "parameters": {
      "CsrfToken": {
        "description": "Token bound to the current opaque session.",
        "in": "header",
        "name": "X-CSRF-Token",
        "required": true,
        "schema": {
          "minLength": 1,
          "type": "string"
        }
      },
      "IdempotencyKey": {
        "in": "header",
        "name": "Idempotency-Key",
        "required": false,
        "schema": {
          "maxLength": 128,
          "minLength": 1,
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Permission, assignment or CSRF validation failed.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
...
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Request rejected.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
...
      "Unauthorized": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "A valid opaque session cookie is required.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      }
    },
    "schemas": {
...
      "CatalogPackage": {
        "additionalProperties": false,
        "properties": {
          "archive_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "type": "string"
          },
          "archive_size_bytes": {
            "maximum": 16777216,
            "minimum": 1,
            "type": "integer"
          },
          "compatibility_status": {
            "const": "unverified",
            "description": "Generic anonymous SteamCMD support remains best effort until an OS/architecture smoke test succeeds."
          },
          "content_size_bytes": {
            "maximum": 33554432,
            "minimum": 1,
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/CatalogFile"
            },
            "maxItems": 63,
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "steam_profile",
              "theme"
            ]
          },
          "manifest": {
            "$ref": "#/components/schemas/CatalogManifest"
          },
          "name": {
            "type": "string"
          },
          "revision": {
            "minimum": 1,
            "type": "integer"
          },
          "schema_version": {
            "const": 1
          },
          "theme_tokens": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ThemeTokens"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "revision",
          "kind",
          "schema_version",
          "name",
          "description",
          "archive_sha256",
          "archive_size_bytes",
          "content_size_bytes",
          "manifest",
          "files",
          "theme_tokens",
          "compatibility_status",
          "created_at"
        ],
        "type": "object"
      },
...
      "Job": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "error_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "instance_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "interaction": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JobInteraction"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string"
          },
          "progress": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer"
          },
          "requested_by": {
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "waiting_for_user",
              "succeeded",
              "failed",
              "cancelled",
              "interrupted"
            ]
          }
        },
        "required": [
          "id",
          "instance_id",
          "kind",
          "state",
          "progress",
          "requested_by",
          "error_code",
          "error_message",
          "created_at",
          "started_at",
          "finished_at",
          "interaction"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "Opaque HttpOnly SameSite=Strict session cookie. Secure is mandatory outside loopback development.",
        "in": "cookie",
        "name": "dmx_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-host game server management API. Secret values are write-only.",
    "title": "DmxServerManager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/catalog": {
      "get": {
        "operationId": "listCatalogPackages",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "enum": [
                "steam_profile",
                "theme"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogPackage"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Latest installed revision of each local package."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/import": {
      "post": {
        "description": "Owner-only validated local .dmxpack import. The package never carries executable code, arbitrary CSS, shell commands or download URLs.",
        "operationId": "importCatalogPackage",
        "parameters": [
          {
            "in": "header",
            "name": "X-Dmx-Package-Sha256",
            "required": true,
            "schema": {
              "pattern": "^[0-9A-Fa-f]{64}$",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "content": {
            "application/vnd.dmxpack+zip": {
              "schema": {
                "format": "binary",
                "maxLength": 16777216,
                "type": "string"
              }
            },
            "application/zip": {
              "schema": {
                "format": "binary",
                "maxLength": 16777216,
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Package staged and persistent import job queued."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/theme": {
      "get": {
        "description": "Returns only the validated closed design-token set and local checksum-pinned PNG assets for the exact active revision. No manifest, CSS, HTML or JavaScript is exposed.",
        "operationId": "getActiveCatalogTheme",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveTheme"
                }
              }
            },
            "description": "Global active theme, or the built-in default.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "catalog"
        ]
      },
      "put": {
        "description": "Owner only. Selects one exact installed theme revision or restores the built-in default.",
        "operationId": "selectCatalogTheme",
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveTheme"
                }
              }
            },
            "description": "Global theme selection updated.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/{kind}/{id}/revisions": {
      "get": {
        "operationId": "listCatalogRevisions",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "steam_profile",
                "theme"
              ]
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 7,
              "pattern": "^(?:steam|theme)-[a-z0-9]+(?:-[a-z0-9]+)*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogPackage"
                  },
                  "type": "array"
                }
              }
            },
            "description": "All active immutable revisions for the package."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/{kind}/{id}/revisions/{revision}": {
      "delete": {
        "description": "Owner-only deletion. A revision pinned by an instance cannot be deleted and its revision number can never be reused.",
        "operationId": "deleteCatalogRevision",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "steam_profile",
                "theme"
              ]
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 7,
              "pattern": "^(?:steam|theme)-[a-z0-9]+(?:-[a-z0-9]+)*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Unused package revision deleted and tombstoned."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "catalog"
        ]
      },
      "get": {
        "operationId": "getCatalogRevision",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "steam_profile",
                "theme"
              ]
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 7,
              "pattern": "^(?:steam|theme)-[a-z0-9]+(?:-[a-z0-9]+)*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogPackage"
                }
              }
            },
            "description": "One immutable installed package revision."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/{kind}/{id}/revisions/{revision}/assets/{asset}": {
      "get": {
        "operationId": "getCatalogAsset",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "steam_profile",
                "theme"
              ]
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 7,
              "pattern": "^(?:steam|theme)-[a-z0-9]+(?:-[a-z0-9]+)*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset",
            "required": true,
            "schema": {
              "enum": [
                "icon",
                "logo",
                "preview"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Checksum-verified PNG asset declared by the package manifest.",
            "headers": {
              "ETag": {
                "description": "Quoted SHA-256 of the immutable asset.",
                "schema": {
                  "pattern": "^\\\"[0-9a-f]{64}\\\"$",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}